Qt6 inline qHash function require size_t instead unit

+++ src/lib/marble/geodata/data/GeoDataCoordinates.h
@@ -431,7 +431,7 @@ class GEODATA_EXPORT GeoDataCoordinates
     static const GeoDataCoordinates null;
 };
 
-GEODATA_EXPORT uint qHash(const GeoDataCoordinates& coordinates );
+GEODATA_EXPORT size_t qHash(const GeoDataCoordinates& coordinates );
 
 
 }
